test: move-only download_from_url to stand-alone util file
Can be reviewed via --color-moved=dimmed-zebra
